Added much better precision to Coordinates

+++ Shapes/Shape.py
@@ -1,8 +1,7 @@
 import math
 from Coordinate import Coordinate, Coordinates
 from tqdm import tqdm
-from CuringCalculations import curing_calculations
-
+from decimal import Decimal
 
 class Shape:
     """
@@ -87,7 +86,7 @@ class Shape:
             end_point = line_coordinates[i + 1]
             num_points = int(self.distance(start_point.x, start_point.y, end_point.x, end_point.y) * (1 / resolution)) + 1
             for j in range(1, num_points):
-                t = j / (num_points - 1)
+                t = Decimal(j / (num_points - 1))
                 x = start_point.x + t * (end_point.x - start_point.x)
                 y = start_point.y + t * (end_point.y - start_point.y)
                 new_coord = Coordinate(x, y)
